Added two axis support for camera slope finder
the measurements are only accurate when the other is at 0.
I'm still trying to figure out how to make them function independtely of
each other, but as long as the phone is held vertically and not tipped forward
or back then you can measure yaw (sidways tilt). Similarly, the closer yaw is to 0
to more accurate the pitch measurement, forward and back will be.

I'm less than pleased with the current behavior.
I went down this path because the demo video was confusing,
my best guess is that it would show you the higher of the two slopes
(between pitch and yaw) but there was no visual indicator saying which one
it was picking for you, My guess is that the user had similar issues as this app.
if they wanted reliable data they would have to choose either pitch or yaw.