A better formula for the modified pitch for vertical view (tilt)
pitch and yaw are both indepent of roll but not each other. Still,
the new formula for pitch is much better than my last one which was
using gravity poorly to deal with some issues and it would make the
slider jump when yaw was affecting it. TL;DR; better algorithm, same
core problem.